refactored sso search methods to fix tech debt - sso search now has own layer files and removed dead code from tms.* layer files, updated unit tests for new layers
