+++ backend/src/tests/tms.test.ts
@@ -1,6 +1,5 @@
 import request from 'supertest'
 import express, { type ErrorRequestHandler } from 'express'
-import axios from 'axios'
 import { TMSRepository } from '../repositories/tms.repository'
 import { groupRepository } from '../repositories/group.repository'
 import { TMSConstants } from '../common/tms.constants'
@@ -2817,648 +2816,6 @@ describe('Tenant API', () => {
         httpResponseCode: 400,
         message: 'Tenant user cannot be removed',
         name: 'Error occurred removing tenant user',
-      })
-    })
-  })
-
-  describe('GET /v1/users/bcgovssousers/idir/search', () => {
-    beforeEach(() => {
-      app.get(
-        '/v1/users/bcgovssousers/idir/search',
-        validate(validator.searchBCGOVSSOUsers, {}, {}),
-        (req, res) => tmsController.searchBCGOVSSOUsers(req, res),
-      )
-
-      const validationErrorHandler: ErrorRequestHandler = (
-        err,
-        req,
-        res,
-        next,
-      ) => {
-        if (
-          err &&
-          typeof err === 'object' &&
-          'name' in err &&
-          (err as { name: string }).name === 'ValidationError'
-        ) {
-          return res
-            .status((err as { statusCode: number }).statusCode)
-            .json(err)
-        }
-        next(err)
-      }
-      app.use(validationErrorHandler)
-    })
-
-    it('should search IDIR users by email successfully', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            guid: 'F45AFBBD68C51D6F956BA3A1DE1878A1',
-            firstName: 'John',
-            lastName: 'Doe',
-            email: 'john.doe@gov.bc.ca',
-            displayName: 'Doe, John',
-          },
-        ],
-      }
-
-      const mockToken = 'mock-access-token'
-      const mockAxiosGet = jest
-        .fn()
-        .mockResolvedValue({ data: mockSearchResults })
-      const mockAxiosPost = jest
-        .fn()
-        .mockResolvedValue({ data: { access_token: mockToken } })
-
-      jest.doMock('axios', () => ({
-        default: {
-          get: mockAxiosGet,
-          post: mockAxiosPost,
-        },
-      }))
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'john.doe@gov.bc.ca' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should search IDIR users by firstName successfully', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            guid: 'F45AFBBD68C51D6F956BA3A1DE1878A2',
-            firstName: 'Jane',
-            lastName: 'Smith',
-            email: 'jane.smith@gov.bc.ca',
-            displayName: 'Smith, Jane',
-          },
-        ],
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ firstName: 'Jane' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should search IDIR users by lastName successfully', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            guid: 'F45AFBBD68C51D6F956BA3A1DE1878A3',
-            firstName: 'Bob',
-            lastName: 'Johnson',
-            email: 'bob.johnson@gov.bc.ca',
-            displayName: 'Johnson, Bob',
-          },
-        ],
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ lastName: 'Johnson' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should search IDIR users by guid successfully', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            guid: 'F45AFBBD68C51D6F956BA3A1DE1878A4',
-            firstName: 'Alice',
-            lastName: 'Williams',
-            email: 'alice.williams@gov.bc.ca',
-            displayName: 'Williams, Alice',
-          },
-        ],
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ guid: 'F45AFBBD68C51D6F956BA3A1DE1878A4' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should deduplicate IDIR users by guid by default', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            firstName: 'John',
-            lastName: '',
-            email: '',
-            username: 'guid-1@idir',
-            attributes: {
-              idir_user_guid: ['GUID-1'],
-              idir_username: ['JDOE'],
-              display_name: ['Doe, John'],
-            },
-          },
-          {
-            firstName: 'John',
-            lastName: 'Doe',
-            email: 'john.doe@gov.bc.ca',
-            username: 'guid-1@idir',
-            attributes: {
-              idir_user_guid: ['GUID-1'],
-              idir_username: ['JDOE'],
-              display_name: ['Doe, John'],
-            },
-          },
-          {
-            firstName: 'Jane',
-            lastName: 'Smith',
-            email: 'jane.smith@gov.bc.ca',
-            username: 'guid-2@idir',
-            attributes: {
-              idir_user_guid: ['GUID-2'],
-              idir_username: ['JSMITH'],
-              display_name: ['Smith, Jane'],
-            },
-          },
-        ],
-      }
-
-      jest
-        .spyOn(
-          tmsController.tmsService as unknown as {
-            getToken: () => Promise<string>
-          },
-          'getToken',
-        )
-        .mockResolvedValue('mock-access-token')
-      jest.spyOn(axios, 'get').mockResolvedValue({ data: mockSearchResults })
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'john.doe@gov.bc.ca' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual({
-        data: [mockSearchResults.data[1], mockSearchResults.data[2]],
-      })
-      expect(axios.get).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({
-          params: { email: 'john.doe@gov.bc.ca' },
-        }),
-      )
-    })
-
-    it('should return empty array when no users found', async () => {
-      const mockSearchResults = { data: [] }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'nonexistent@gov.bc.ca' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should return 400 when no search parameters provided', async () => {
-      const response = await request(app).get(
-        '/v1/users/bcgovssousers/idir/search',
-      )
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when firstName is too short', async () => {
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ firstName: 'A' })
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when dedup query parameter is provided', async () => {
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'john.doe@gov.bc.ca', dedup: 'true' })
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when lastName is too short', async () => {
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ lastName: 'B' })
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when BC GOV SSO API returns bad request', async () => {
-      const error = {
-        response: {
-          status: 400,
-          data: { message: 'Invalid search parameters' },
-        },
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockRejectedValue(
-          new BadRequestError(
-            `BC GOV SSO API returned bad request: ${error.response.data.message}`,
-          ),
-        )
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'invalid@email' })
-
-      expect(response.status).toBe(400)
-      expect(response.body).toMatchObject({
-        errorMessage: 'Bad Request',
-        httpResponseCode: 400,
-        message: expect.stringContaining('BC GOV SSO API returned bad request'),
-        name: 'Error occurred searching SSO users',
-      })
-    })
-
-    it('should return 500 when API error occurs', async () => {
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOUsers')
-        .mockRejectedValue(new Error('Error invoking BC GOV SSO API'))
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/idir/search')
-        .query({ email: 'test@gov.bc.ca' })
-
-      expect(response.status).toBe(500)
-      expect(response.body).toMatchObject({
-        errorMessage: 'Internal Server Error',
-        httpResponseCode: 500,
-        message: 'Error invoking BC GOV SSO API',
-        name: 'Error occurred searching SSO users',
-      })
-    })
-  })
-
-  describe('GET /v1/users/bcgovssousers/bceid/search', () => {
-    beforeEach(() => {
-      app.get(
-        '/v1/users/bcgovssousers/bceid/search',
-        validate(validator.searchBCGOVSSOBceidUsers, {}, {}),
-        (req, res) => tmsController.searchBCGOVSSOBceidUsers(req, res),
-      )
-
-      const validationErrorHandler: ErrorRequestHandler = (
-        err,
-        req,
-        res,
-        next,
-      ) => {
-        if (
-          err &&
-          typeof err === 'object' &&
-          'name' in err &&
-          (err as { name: string }).name === 'ValidationError'
-        ) {
-          return res
-            .status((err as { statusCode: number }).statusCode)
-            .json(err)
-        }
-        next(err)
-      }
-      app.use(validationErrorHandler)
-    })
-
-    it('should search BCEID users by displayName successfully', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            guid: 'F45AFBBD68C51D6F956BA3A1DE1878B2',
-            displayName: 'Business User',
-            username: 'businessuser',
-            bceidType: 'business',
-          },
-        ],
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOBceidUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({ bceidType: 'business', displayName: 'Business User' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should search BCEID users with business type before filtering to business users', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            firstName: 'Business',
-            lastName: 'User',
-            email: 'business.user@example.com',
-            username: 'business.user@bceidbusiness',
-            attributes: {
-              bceid_user_guid: ['GUID-1'],
-              bceid_username: ['BUSINESSUSER'],
-              display_name: ['Business User'],
-              bceid_business_guid: ['BUSINESS-GUID-1'],
-            },
-          },
-        ],
-      }
-
-      jest
-        .spyOn(
-          tmsController.tmsService as unknown as {
-            getToken: () => Promise<string>
-          },
-          'getToken',
-        )
-        .mockResolvedValue('mock-access-token')
-      jest.spyOn(axios, 'get').mockResolvedValue({ data: mockSearchResults })
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'business',
-          username: 'business.user',
-        })
-
-      expect(response.status).toBe(200)
-      expect(axios.get).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({
-          params: { bceidType: 'business', username: 'business.user' },
-        }),
-      )
-    })
-
-    it('should accept both BCEID search type and filter out basic BCEID users', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            firstName: 'Basic',
-            lastName: 'User',
-            email: 'basic.user@example.com',
-            username: 'basic.user@bceidboth',
-            attributes: {
-              bceid_user_guid: ['BASIC-GUID-1'],
-              bceid_username: ['BASICUSER'],
-              display_name: ['Basic User'],
-            },
-          },
-          {
-            firstName: 'Business',
-            lastName: 'User',
-            email: 'business.user@example.com',
-            username: 'business.user@bceidboth',
-            attributes: {
-              bceid_user_guid: ['BUSINESS-USER-GUID-1'],
-              bceid_username: ['BUSINESSUSER'],
-              display_name: ['Business User'],
-              bceid_business_guid: ['BUSINESS-GUID-1'],
-            },
-          },
-        ],
-      }
-
-      jest
-        .spyOn(
-          tmsController.tmsService as unknown as {
-            getToken: () => Promise<string>
-          },
-          'getToken',
-        )
-        .mockResolvedValue('mock-access-token')
-      jest.spyOn(axios, 'get').mockResolvedValue({ data: mockSearchResults })
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'both',
-          username: 'business.user',
-        })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual({
-        data: [mockSearchResults.data[1]],
-      })
-      expect(axios.get).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({
-          params: { bceidType: 'business', username: 'business.user' },
-        }),
-      )
-    })
-
-    it('should deduplicate BCEID users by guid by default', async () => {
-      const mockSearchResults = {
-        data: [
-          {
-            firstName: 'Master',
-            lastName: '',
-            email: '',
-            username: 'guid-1@bceidbusiness',
-            attributes: {
-              bceid_user_guid: ['GUID-1'],
-              bceid_username: ['USER1'],
-              display_name: ['Master Chief'],
-              bceid_business_guid: ['BUSINESS-GUID-1'],
-            },
-          },
-          {
-            firstName: 'Master',
-            lastName: 'Chief',
-            email: 'master.chief@example.com',
-            username: 'guid-1@bceidbusiness',
-            attributes: {
-              bceid_user_guid: ['GUID-1'],
-              bceid_username: ['USER1'],
-              display_name: ['Master Chief'],
-              bceid_business_guid: ['BUSINESS-GUID-1'],
-            },
-          },
-          {
-            firstName: 'Shankar',
-            lastName: 'Sethuraman',
-            email: 'shankar@example.com',
-            username: 'guid-2@bceidbusiness',
-            attributes: {
-              bceid_user_guid: ['GUID-2'],
-              bceid_username: ['USER2'],
-              display_name: ['Shankar Sethuraman'],
-              bceid_business_guid: ['BUSINESS-GUID-2'],
-            },
-          },
-        ],
-      }
-
-      jest
-        .spyOn(
-          tmsController.tmsService as unknown as {
-            getToken: () => Promise<string>
-          },
-          'getToken',
-        )
-        .mockResolvedValue('mock-access-token')
-      jest.spyOn(axios, 'get').mockResolvedValue({ data: mockSearchResults })
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'business',
-          username: 'user1',
-        })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual({
-        data: [mockSearchResults.data[1], mockSearchResults.data[2]],
-      })
-      expect(axios.get).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({
-          params: { bceidType: 'business', username: 'user1' },
-        }),
-      )
-    })
-
-    it('should return empty array when no users found', async () => {
-      const mockSearchResults = { data: [] }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOBceidUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({ bceidType: 'business', guid: 'NONEXISTENT' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should default to business when bceidType is missing', async () => {
-      const mockSearchResults = { data: [] }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOBceidUsers')
-        .mockResolvedValue(mockSearchResults as unknown)
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({ guid: 'F45AFBBD68C51D6F956BA3A1DE1878B1' })
-
-      expect(response.status).toBe(200)
-      expect(response.body).toEqual(mockSearchResults)
-    })
-
-    it('should return 400 when bceidType is invalid', async () => {
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'invalid',
-          guid: 'F45AFBBD68C51D6F956BA3A1DE1878B1',
-        })
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when dedup query parameter is provided', async () => {
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'business',
-          guid: 'F45AFBBD68C51D6F956BA3A1DE1878B1',
-          dedup: 'true',
-        })
-
-      expect(response.status).toBe(400)
-      expect(response.body.message).toBe('Validation Failed')
-    })
-
-    it('should return 400 when BC GOV SSO BCEID API returns bad request', async () => {
-      const error = {
-        response: {
-          status: 400,
-          data: { message: 'Invalid search parameters' },
-        },
-      }
-
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOBceidUsers')
-        .mockRejectedValue(
-          new BadRequestError(
-            `BC GOV SSO BCEID API returned bad request: ${error.response.data.message}`,
-          ),
-        )
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({ bceidType: 'business', guid: 'invalid' })
-
-      expect(response.status).toBe(400)
-      expect(response.body).toMatchObject({
-        errorMessage: 'Bad Request',
-        httpResponseCode: 400,
-        message: expect.stringContaining(
-          'BC GOV SSO BCEID API returned bad request',
-        ),
-        name: 'Error occurred searching BCEID users',
-      })
-    })
-
-    it('should return 500 when API error occurs', async () => {
-      jest
-        .spyOn(tmsController.tmsService, 'searchBCGOVSSOBceidUsers')
-        .mockRejectedValue(new Error('Error invoking BC GOV SSO BCEID API'))
-
-      const response = await request(app)
-        .get('/v1/users/bcgovssousers/bceid/search')
-        .query({
-          bceidType: 'business',
-          guid: 'F45AFBBD68C51D6F956BA3A1DE1878B1',
-        })
-
-      expect(response.status).toBe(500)
-      expect(response.body).toMatchObject({
-        errorMessage: 'Internal Server Error',
-        httpResponseCode: 500,
-        message: 'Error invoking BC GOV SSO BCEID API',
-        name: 'Error occurred searching BCEID users',
       })
     })
   })
